+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.1.0
+
+- fix: `/list` and `/:param` are not separate routes, `list` will be resolved first,
+if states sooner
+- feat: handler type changed, now it accepts `Res` instead of `Req`, which allows
+access to `params`, `route` and so on
+- fix: `GET` method is now default
+
 # 2.0.1
 
 - fix: one endpoint with different methods